bnxt_re/lib: Resize CQ support
Adds resize CQ support. After issuing the resize_cq cmd, lib
reaps all the existing completions in the CQ and store it in
a list. The reaping continues till the final CQE. Once final
CQE is seen, it issues a ibv_cmd_poll_cq so the poll_cq context
in the kernel is invoked and driver frees up the previous CQ
resources.

During a poll CQ from application, the pending completions in
the list is completed first before polling completions from the
new CQ.

Signed-off-by: Selvin Xavier <selvin.xavier@broadcom.com>